#11-modify-테스트디렉토리 구조 변경
